API

CTProblems.CTProblems — Module

CTProblems module.

Lists all the imported modules and packages:

Base

CTBase

CTFlows

Core

DocStringExtensions

ForwardDiff

LinearAlgebra

MINPACK

OrdinaryDiffEq

List of all the exported names:

Problem

@Problems

Problems

@ProblemsList

ProblemsList

plot

CTProblems.Problem — Method

Problem(description...) -> Any

Returns the optimal control problem described by description. See the Introduction and page list of problems for details.

Example

julia> Problem(:integrator, :energy)

CTProblems.Problems — Method

Problems(
    description...
) -> Tuple{Vararg{CTProblems.OptimalControlProblem}}

Returns the list of problems consistent with the description, as a Tuple of OptimalControlProblem, see the page list of problems for details.

Example

julia> Problems(:integrator, :energy)

CTProblems.Problems — Method

Problems(
    expr::Union{Expr, QuoteNode}
) -> Tuple{Vararg{CTProblems.OptimalControlProblem}}

Returns the list of problems consistent with the expression, as a Tuple of OptimalControlProblem, see the page list of problems for details.

Example

julia> Problems(:(:integrator & :energy))

See also

@Problems for a more natural usage.

Note

You have to define a logical condition with the combination of symbols and the three operators: !, | and &, respectively for the negation, the disjunction and the conjunction.

CTProblems.ProblemsList — Method

ProblemsList(
    description...
) -> Tuple{Vararg{Tuple{Vararg{Symbol}}}}

Returns the list of problems descriptions consistent with the description, as a Tuple of Description, see the page list of problems descriptions for details.

Example

julia> ProblemsList(:integrator, :energy)

CTProblems.ProblemsList — Method

ProblemsList(
    expr::Union{Expr, QuoteNode}
) -> Tuple{Vararg{Tuple{Vararg{Symbol}}}}

Returns the list of problems descriptions consistent with the expression, as a Tuple of Description, see the list of problems descriptions page for details.

Example

julia> ProblemsList(:(:integrator & :energy))

See also

@ProblemsList for a more natural usage.

Note

You have to define a logical condition with the combination of symbols and the three operators: !, | and &, respectively for the negation, the disjunction and the conjunction.

RecipesBase.plot — Method

plot(sol; kwargs...) -> Plots.Plot

A binding to CTBase.jl plot function.

CTProblems.@Problems — Macro

Returns the list of problems consistent with the description, as a Tuple of OptimalControlProblem, see the page list of problems for details.

Example

julia> @Problems :integrator & :energy

Note

You have to define a logical condition with the combination of symbols and the three operators: !, | and &, respectively for the negation, the disjunction and the conjunction.

CTProblems.@ProblemsList — Macro

Returns the list of problems descriptions consistent with the expression, as a Tuple of Description, see the list of problems descriptions page for details.

Example

julia> @ProblemsList :integrator & :energy

Note

You have to define a logical condition with the combination of symbols and the three operators: !, | and &, respectively for the negation, the disjunction and the conjunction.